refactor(MCP): migrate http_client server to new metadata pattern
Move http_client server from single-file implementation to folder
structure with createToolsRecord/buildTools pattern:
- Add metadata.ts with HTTP_CLIENT_SERVER, HTTP_CLIENT_TOOLS_METADATA
- Add tools/index.ts with send_request handler
- Add index.ts with createServer that registers http_client + web tools
- Update constants.ts to use metadata: HTTP_CLIENT_SERVER
- Update servers/index.ts import path
- Delete old single-file implementation
